Refactof: 모달 컴포넌트 코드 정리
모달 컴포넌트 코드 정리
