+++ src/components/header/header.jsx
@@ -6,8 +6,8 @@ import IconArrowLeft from '../../assets/icon/icon-arrow-left.png'
 import IconLuckitLogo from '../../assets/icon/icon-luckit.png'
 import IconMoreVertical from '../../assets/icon/icon- more-vertical.png'
 import { StoreBtn, UploadBtn } from '../../components/button/button'
-import {PostModalWrap, PostModalBtnWrap, NavLinkStyle, ModalBtnWrap, ModalWrap, ModalNavLink, Div} from '../../components/modal/modalstyle'
-// import { LogoutModal, PostModal } from '../modal/modal'
+import { ChatRoomModal, LogoutModal } from '../modal/modal'
+
 
 
 // 팔로잉 없을 때 홈페이지, 팔로잉 있을 때 홈페이지(스크롤X)
@@ -50,58 +50,29 @@ export const SearchHeader = ({value, onChange}) => {
 // 프로필, 채팅 페이지
 export const ProfileAndChatHeader = () => {
   const [isOpen, setIsOpen] = useState(false);
-  const [isOpenModal, setIsOpenModal] = useState(false)
   const navigate = useNavigate();
 
   const onClick = () => {
     setIsOpen(true);
   }
 
-  const onClickCloseModal = () => {
-    setIsOpen(false);
-  }
-
-  const onClickLogoutModal = () => {
-    setIsOpenModal(true)
-  }
-
-  const onClickCancel = () => {
-    setIsOpenModal(false)
-  }
-
-  const onClickLogout = () => {
-    localStorage.clear();
+  const onClickClose = (value) => {
+    setIsOpen(value);
   }
 
   return (
-    <ProfileHeaderWrap>
-      <button onClick={() => {navigate(-1)}}>
-        <img src={IconArrowLeft} alt='뒤로가기' />
-      </button>
-      <button onClick={onClick}>
-        <img src={IconMoreVertical} alt='설정 및 로그아웃 버튼' />
-      </button>
-      {isOpen && 
-      <PostModalWrap onClick={onClickCloseModal}>
-        <div className='test' onClick={(e) => e.stopPropagation()}>
-          <PostModalBtnWrap>
-            <NavLinkStyle to='/editprofile'>설정 및 개인 정보</NavLinkStyle>
-            <button onClick={onClickLogoutModal}>로그아웃</button>
-          </PostModalBtnWrap>
-          {isOpenModal && 
-          <Div>
-            <ModalWrap>
-              <strong>로그아웃 하시겠습니까?</strong>
-              <ModalBtnWrap>
-                <button onClick={onClickCancel}>취소</button>
-                <ModalNavLink to='/login' onClick={onClickLogout}>로그아웃</ModalNavLink>
-              </ModalBtnWrap>
-            </ModalWrap>
-          </Div>
-          }
-        </div>
-      </PostModalWrap>}
-    </ProfileHeaderWrap>
+    <>
+      <ProfileHeaderWrap>
+        <button onClick={() => {navigate(-1)}}>
+          <img src={IconArrowLeft} alt='뒤로가기' />
+        </button>
+        <button onClick={onClick}>
+          <img src={IconMoreVertical} alt='설정 및 로그아웃 버튼' />
+        </button>
+      </ProfileHeaderWrap>
+        {isOpen && 
+        <LogoutModal onClickClose={onClickClose}/>}
+    </>
   )
 }
 
@@ -170,8 +141,8 @@ export const ChatRoomHeader = () => {
     setIsOpen(true);
   }
 
-  const onClickCloseModal = () => {
-    setIsOpen(false);
+  const onClickClose = (value) => {
+    setIsOpen(value);
   }
 
   return (
@@ -183,14 +154,7 @@ export const ChatRoomHeader = () => {
       <button onClick={onClick}>
         <img src={IconMoreVertical} alt='채팅방 나가기 버튼' />
       </button>
-      {isOpen && 
-      <PostModalWrap onClick={onClickCloseModal}>
-        <div className='test' onClick={(e) => e.stopPropagation()}>
-          <PostModalBtnWrap>
-            <NavLinkStyle to='/chatpage'>채팅방 나가기</NavLinkStyle>
-          </PostModalBtnWrap>
-        </div>
-      </PostModalWrap>}
+      {isOpen && <ChatRoomModal onClickClose={onClickClose}/>}
     </ProfileHeaderWrap>
   )
 }
